+++ packages/alert/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pikas-ui/alert
 
+## 2.0.9
+
+### Patch Changes
+
+- Update dependencies
+- Updated dependencies
+  - @pikas-ui/icons@2.0.9
+  - @pikas-ui/styles@2.1.7
+
 ## 2.0.8
 
 ### Patch Changes