+++ packages/dropdown-menu/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pikas-ui/dropdown-menu
 
+## 2.0.9
+
+### Patch Changes
+
+- Update dependencies
+- Updated dependencies
+  - @pikas-ui/icons@2.0.9
+  - @pikas-ui/loader@2.0.8
+  - @pikas-ui/menu@2.0.9
+  - @pikas-ui/styles@2.1.7
+
 ## 2.0.8
 
 ### Patch Changes