+++ packages/searchbar/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pikas-ui/searchbar
 
+## 2.0.11
+
+### Patch Changes
+
+- Update dependencies
+- Updated dependencies
+- Updated dependencies
+  - @pikas-ui/button@2.0.10
+  - @pikas-ui/icons@2.0.9
+  - @pikas-ui/loader@2.0.8
+  - @pikas-ui/separator@2.0.8
+  - @pikas-ui/styles@2.1.7
+  - @pikas-ui/textfield@2.0.10
+
 ## 2.0.10
 
 ### Patch Changes