+++ packages/separator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pikas-ui/separator
 
+## 2.0.8
+
+### Patch Changes
+
+- Update dependencies
+- Updated dependencies
+  - @pikas-ui/styles@2.1.7
+
 ## 2.0.7
 
 ### Patch Changes