+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pikas-ui/styles
 
+## 2.1.7
+
+### Patch Changes
+
+- Update dependencies
+
 ## 2.1.6
 
 ### Patch Changes